fix test problems

+++ src/main/java/org/apache/iotdb/quality/dprofile/UDTFMvAvg.java
@@ -42,7 +42,7 @@ public class UDTFMvAvg implements UDTF {
         .validate(
             x -> (int) x > 0,
             "Window size should be larger than 0.",
-            validator.getParameters().getDoubleOrDefault("window", 10));
+            validator.getParameters().getIntOrDefault("window", 10));
   }
 
   @Override
